+++ src/components/index.ts
@@ -1,3 +1,4 @@
+export * from './grid/grid.component';
 export * from './tableGridColumn/tableGridColumn.component';
 export * from './tableGridMetadataGatherer/tableGridMetadataGatherer.component';
 export * from './tableGridMetadataGatherer/tableGridMetadataGatherer.interface';