+++ src/components/metadata/components.ts
@@ -1,5 +1,0 @@
-export * from '../tableGridMetadataGatherer/tableGridMetadataGatherer.component';
-export * from '../tableGridColumn/tableGridColumn.component';
-export * from './groupedTable/groupedTableMetadataGatherer.component';
-export * from './groupedTable/headerTableGroup.component';
-export * from './groupedTable/headerTableGroupColumn.component';