+++ src/components/metadata/index.ts
@@ -1,2 +1,0 @@
-export * from '../tableGridMetadataGatherer/tableGridMetadataGatherer.interface';
-export * from './groupedTable/groupedTable.interface';