+++ src/components/metadata/types.ts
@@ -1,9 +1,0 @@
-/**
- * Token for getting component that is used as metadata gatherer
- */
-export class METADATA_GATHERER
-{
-}
-
-export * from '../tableGridColumn/tableGridColumn.context';
-export * from './groupedTable/headerTableGroup.context';