+++ src/interfaces/gridOptions/gridOptions.interface.ts
@@ -1,6 +1,3 @@
-import {PagingPosition} from './enums';
-import {GridPluginTypes} from './types';
-
 /**
  * Describes grid options used for grid
  */