+++ src/interfaces/gridPluginInstances/gridPluginInstances.interface.ts
@@ -1,0 +1,7 @@
+/**
+ * Interface describing object storing all existing plugin instances for grid
+ */
+export interface GridPluginInstances
+{
+    [pluginName: string]: GridPlugin;
+}