+++ src/interfaces/pluginOptions/pluginOptions.interface.ts
@@ -1,0 +1,6 @@
+/**
+ * Base options for every plugin
+ */
+export interface PluginOptions
+{
+}