+++ src/misc/index.ts
@@ -1,2 +1,0 @@
-export * from './plugin.interface';
-export * from './gridOptions.interface';